Move `pythonDefRun()` into pythonFileUtils.ts to live with its friends.
